move format method into format util class
